Added Suspense boundary for navbar
