+++ apps/site/src/app/(main)/layout.tsx
@@ -1,8 +1,9 @@
-import { PropsWithChildren } from "react";
+import { PropsWithChildren, Suspense } from "react";
 
 import type { Metadata } from "next";
 
 import Footer from "@/lib/components/Footer/Footer";
+import BaseNavbar from "@/lib/components/Navbar/BaseNavbar";
 import NavbarParent from "@/lib/components/Navbar/NavbarParent";
 import stars from "@/assets/backgrounds/starry_repeatable.png";
 
@@ -20,7 +21,9 @@ export default function Layout({ children }: PropsWithChildren) {
 			style={{ backgroundImage: `url(${stars.src})` }}
 			className="overflow-x-hidden bg-top bg-repeat-y bg-[length:100%]"
 		>
-			<NavbarParent />
+			<Suspense fallback={<BaseNavbar />}>
+				<NavbarParent />
+			</Suspense>
 			{children}
 			<Footer />
 		</div>
